improve folder structure
