+++ src/resolvers/index.ts
@@ -1,5 +1,4 @@
 import { ChannelFollowsInfosResolver } from './followsInfo'
 import { VideoViewsInfosResolver } from './viewsInfo'
-import { queryNodeStitchingResolvers } from './queryNodeStitchingResolvers'
 
-export { ChannelFollowsInfosResolver, VideoViewsInfosResolver, queryNodeStitchingResolvers }
+export { ChannelFollowsInfosResolver, VideoViewsInfosResolver }
